Hangsuu

+++ insider/src/main/java/com/kh/insider/dto/BoardDto.java
@@ -12,8 +12,8 @@ public class BoardDto {
 	public long memberNo;
 	public String boardContent;
 	public int boardLike;
-	public int boardLat;
-	public int boardLon;
+	public double boardLat;
+	public double boardLon;
 	public Date boardTime;
 	public int boardReply;
 	public int boardReplyValid;
